Refactor incubating clients into separate package.
In the interest of completely separating the production and incubating
namespaces, I created an incubating package. For clarity, the modules,
names, and tests parallel those in the the rest of the SDK.

I also separated utility functions into a separate module.
