+++ src/momento/incubating/__init__.py
@@ -1,0 +1,1 @@
+INCUBATING_WARNING_MSG = "Using the incubating client: functionality and features are experimental and subject to change or deletion!"
